move dynamic charset wrapper out of contrib
