+++ docs/contrib/wrappers.rst
@@ -12,6 +12,10 @@ Extra Wrappers
 
 .. autoclass:: JSONRequestMixin
 
+.. autoclass:: DynamicCharsetRequestMixin
+
+.. autoclass:: DynamicCharsetResponseMixin
+
 .. autoclass:: ProtobufRequestMixin
    :members:
 
@@ -19,10 +23,4 @@ Extra Wrappers
    :members:
 
 .. autoclass:: ReverseSlashBehaviorRequestMixin
-   :members:
-
-.. autoclass:: DynamicCharsetRequestMixin
-   :members:
-
-.. autoclass:: DynamicCharsetResponseMixin
    :members:
