+++ tests/contrib/test_wrappers.py
@@ -11,7 +11,7 @@
 
 from werkzeug.contrib import wrappers
 from werkzeug import routing
-from werkzeug.wrappers import Request, Response
+from werkzeug.wrappers import Request
 
 
 def test_reverse_slash_behavior():
@@ -29,57 +29,3 @@ def test_reverse_slash_behavior():
     assert adapter.match() == ('foo', {})
     adapter = map.bind(req.host, req.script_root)
     assert adapter.match(req.path) == ('foo', {})
-
-
-def test_dynamic_charset_request_mixin():
-    class MyRequest(wrappers.DynamicCharsetRequestMixin, Request):
-        pass
-    env = {'CONTENT_TYPE': 'text/html'}
-    req = MyRequest(env)
-    assert req.charset == 'latin1'
-
-    env = {'CONTENT_TYPE': 'text/html; charset=utf-8'}
-    req = MyRequest(env)
-    assert req.charset == 'utf-8'
-
-    env = {'CONTENT_TYPE': 'application/octet-stream'}
-    req = MyRequest(env)
-    assert req.charset == 'latin1'
-    assert req.url_charset == 'latin1'
-
-    MyRequest.url_charset = 'utf-8'
-    env = {'CONTENT_TYPE': 'application/octet-stream'}
-    req = MyRequest(env)
-    assert req.charset == 'latin1'
-    assert req.url_charset == 'utf-8'
-
-    def return_ascii(x):
-        return "ascii"
-    env = {'CONTENT_TYPE': 'text/plain; charset=x-weird-charset'}
-    req = MyRequest(env)
-    req.unknown_charset = return_ascii
-    assert req.charset == 'ascii'
-    assert req.url_charset == 'utf-8'
-
-
-def test_dynamic_charset_response_mixin():
-    class MyResponse(wrappers.DynamicCharsetResponseMixin, Response):
-        default_charset = 'utf-7'
-    resp = MyResponse(mimetype='text/html')
-    assert resp.charset == 'utf-7'
-    resp.charset = 'utf-8'
-    assert resp.charset == 'utf-8'
-    assert resp.mimetype == 'text/html'
-    assert resp.mimetype_params == {'charset': 'utf-8'}
-    resp.mimetype_params['charset'] = 'iso-8859-15'
-    assert resp.charset == 'iso-8859-15'
-    resp.set_data(u'Hällo Wörld')
-    assert b''.join(resp.iter_encoded()) == \
-           u'Hällo Wörld'.encode('iso-8859-15')
-    del resp.headers['content-type']
-    try:
-        resp.charset = 'utf-8'
-    except TypeError:
-        pass
-    else:
-        assert False, 'expected type error on charset setting without ct'
