+++ tests/test_wrappers.py
@@ -23,6 +23,10 @@ from tests import strict_eq
 from werkzeug import wrappers
 from werkzeug.exceptions import SecurityError, RequestedRangeNotSatisfiable, \
     BadRequest
+from werkzeug.wrappers.charset import (
+    DynamicCharsetRequestMixin,
+    DynamicCharsetResponseMixin,
+)
 from werkzeug.wrappers.json import JSONMixin
 from werkzeug.wsgi import LimitedStream, wrap_file
 from werkzeug.datastructures import MultiDict, ImmutableOrderedMultiDict, \
@@ -1299,3 +1303,57 @@ def test_json_request_mixin():
         content_type='text/json'
     )
     assert req.json == {u'foä': 'bar'}
+
+
+def test_dynamic_charset_request_mixin():
+    class MyRequest(DynamicCharsetRequestMixin, wrappers.Request):
+        pass
+    env = {'CONTENT_TYPE': 'text/html'}
+    req = MyRequest(env)
+    assert req.charset == 'latin1'
+
+    env = {'CONTENT_TYPE': 'text/html; charset=utf-8'}
+    req = MyRequest(env)
+    assert req.charset == 'utf-8'
+
+    env = {'CONTENT_TYPE': 'application/octet-stream'}
+    req = MyRequest(env)
+    assert req.charset == 'latin1'
+    assert req.url_charset == 'latin1'
+
+    MyRequest.url_charset = 'utf-8'
+    env = {'CONTENT_TYPE': 'application/octet-stream'}
+    req = MyRequest(env)
+    assert req.charset == 'latin1'
+    assert req.url_charset == 'utf-8'
+
+    def return_ascii(x):
+        return "ascii"
+    env = {'CONTENT_TYPE': 'text/plain; charset=x-weird-charset'}
+    req = MyRequest(env)
+    req.unknown_charset = return_ascii
+    assert req.charset == 'ascii'
+    assert req.url_charset == 'utf-8'
+
+
+def test_dynamic_charset_response_mixin():
+    class MyResponse(DynamicCharsetResponseMixin, wrappers.Response):
+        default_charset = 'utf-7'
+    resp = MyResponse(mimetype='text/html')
+    assert resp.charset == 'utf-7'
+    resp.charset = 'utf-8'
+    assert resp.charset == 'utf-8'
+    assert resp.mimetype == 'text/html'
+    assert resp.mimetype_params == {'charset': 'utf-8'}
+    resp.mimetype_params['charset'] = 'iso-8859-15'
+    assert resp.charset == 'iso-8859-15'
+    resp.set_data(u'Hällo Wörld')
+    assert b''.join(resp.iter_encoded()) == \
+           u'Hällo Wörld'.encode('iso-8859-15')
+    del resp.headers['content-type']
+    try:
+        resp.charset = 'utf-8'
+    except TypeError:
+        pass
+    else:
+        assert False, 'expected type error on charset setting without ct'
